Move english word list to be represented by a Vector in memory
