+++ core-test/src/test/scala/org/bitcoins/core/crypto/words/EnglishWordsTest.scala
@@ -1,0 +1,21 @@
+package org.bitcoins.core.crypto.words
+
+import org.bitcoins.testkit.util.BitcoinSUnitTest
+
+import scala.io.Source
+
+class EnglishWordsTest extends BitcoinSUnitTest {
+  behavior of "EnglishWords"
+
+  private val ENGLISH_WORDS_FILE = "/bip39-wordlists/english.txt"
+
+  it must "have the same in memory and file contents" in {
+    val resourceStream = getClass.getResourceAsStream(ENGLISH_WORDS_FILE)
+    val source = Source.fromInputStream(resourceStream)
+
+    val lines = source.getLines()
+    val linesVec = lines.toVector
+    source.close()
+    assert(EnglishWords.getWords == linesVec)
+  }
+}
