+++ core/src/main/scala/org/bitcoins/core/crypto/words/MnemonicWords.scala
@@ -1,0 +1,5 @@
+package org.bitcoins.core.crypto.words
+
+trait MnemonicWords {
+  def getWords: Vector[String]
+}
